==> 2-Data_processing.ipynb <==
# Processing the Data
* Loading Cleaned Data
* Toeknizing
* Saving Processed data

import pandas as pd
import pickle
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.preprocessing import text, sequence

### 1- Loading

#starting data and ending data folders
in_folder='data/2-cleaned_data'
out_folder='data/3-processed_data'

#loading our data
X_train=pd.read_csv(in_folder+'/X_train.csv')
Y_train=pd.read_csv(in_folder+'/Y_train.csv')
X_test=pd.read_csv(in_folder+'/X_test.csv')
Y_test=pd.read_csv(in_folder+'/Y_test.csv')
labels=list(Y_test.columns)

X_train = X_train["comment_text"].values
X_test  = X_test["comment_text"].values

Y_train = Y_train[labels].values
Y_test  = Y_test[labels].values

X_train[0]

### 2- Tokenizing

# tokenizing the data
tokenizer = Tokenizer(num_words=20000)
tokenizer.fit_on_texts(list(X_train))

# turning the tokenized text into sequences
X_train = tokenizer.texts_to_sequences(X_train)
X_test  = tokenizer.texts_to_sequences(X_test)

# padding the sequences
X_train = sequence.pad_sequences(X_train, maxlen=200)
X_test  = sequence.pad_sequences(X_test,  maxlen=200)

print('X_train shape:', X_train.shape)
print('X_test shape: ', X_test.shape)

### 3- Saving Data for Modelling

#Savink Tokenizer with Pickle
with open(out_folder+'/tokenizer.pickle', 'wb') as t:
    pickle.dump(tokenizer, t, protocol=pickle.HIGHEST_PROTOCOL)

#Saving processed data
pd.DataFrame(X_train).to_csv(out_folder+'/X_train.csv',index=False)
pd.DataFrame(Y_train,columns=labels).to_csv(out_folder+'/Y_train.csv',index=False)
pd.DataFrame(X_test).to_csv(out_folder+'/X_test.csv',index=False)
pd.DataFrame(Y_test,columns=labels).to_csv(out_folder+'/Y_test.csv',index=False)
