==> 5-Modelling - Recurrent NN.ipynb <==
# Recurrent Neural Network

import pandas as pd
import itertools
import h5py
import pickle
from keras import initializers, regularizers, constraints, optimizers, layers
from keras.models import Model, Input, Sequential
from keras.layers import Dense, Input, LSTM, Embedding, Dropout, SpatialDropout1D, Activation
from keras.layers import Conv1D, Bidirectional, GlobalMaxPool1D, MaxPooling1D, BatchNormalization
from keras.optimizers import Adam

in_folder='data/3-processed_data'
out_folder='models'

with open(in_folder+'/tokenizer.pickle', 'rb') as t:
    tokenizer = pickle.load(t)

X_train=pd.read_csv(in_folder+'/X_train.csv')
Y_train=pd.read_csv(in_folder+'/Y_train.csv')
X_test=pd.read_csv(in_folder+'/X_test.csv')
Y_test=pd.read_csv(in_folder+'/Y_test.csv')
labels=list(Y_test.columns)

X_train

# number of unique words we want to use (or: number of rows in incoming embedding vector)
max_features = 20000 

# max number of words in a comment to use (or: number of columns in incoming embedding vector)
max_len = 200 

# dimension of the embedding variable (or: number of rows in output of embedding vector)
embedding_dims = 128

# instantiate RNN model
rnn_model = Sequential()

# add embedding layer 
rnn_model.add(Embedding(input_dim=max_features, input_length=max_len,
                        output_dim=embedding_dims))

# set the dropout layer to drop out 50% of the nodes
rnn_model.add(SpatialDropout1D(0.5))

# add bidirectional layer and pass in an LSTM()
rnn_model.add(Bidirectional(LSTM(25, return_sequences=True)))

# add normalization layer
rnn_model.add(BatchNormalization())

# add pooling layer 
rnn_model.add(GlobalMaxPool1D())

# set the dropout layer to drop out 50% of the nodes
rnn_model.add(Dropout(0.5))

# add dense layer to produce an output dimension of 50 and using relu activation
rnn_model.add(Dense(50, activation='relu'))

# finally add a dense layer
rnn_model.add(Dense(7, activation='sigmoid'))

rnn_model.compile(loss='binary_crossentropy',
                  optimizer=Adam(0.01),
                  metrics=['accuracy'])

rnn_model.summary()

rnn_hist = rnn_model.fit(X_train, Y_train, batch_size=128, 
                          epochs=5, validation_split=0.2)

rnn_test_loss, rnn_test_acc = rnn_model.evaluate(X_test, Y_test, batch_size=128)
print('Test Loss:    ', rnn_test_loss)
print('Test Accuracy:', rnn_test_acc)

## Saving the Model

#Saving Model
rnn_model.save(out_folder+'/rnn_model.h5')

#Saving Performance
models_performance=pd.read_csv('models/models_performance.csv')
stats=['Recurrent Model', rnn_test_loss,rnn_test_acc]
models_performance.loc[len(models_performance),:]=stats
models_performance.to_csv(out_folder+'/models_performance.csv',index=False)
