==> 6-Conclusion.ipynb <==
## Conclusion

import pandas as pd
import h5py
from keras.models import load_model
import pickle
import itertools
import numpy as np
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.preprocessing import text, sequence
from nltk.tokenize import word_tokenize
import matplotlib.pyplot as plt

df=pd.read_csv('models/models_performance.csv')
df

**Our best performing model is the Recurrent Neural Network, lets load it and test it!**

#loading Tokenizer
with open('data/3-processed_data/tokenizer.pickle', 'rb') as t:
    tokenizer = pickle.load(t)
#loading test data
Y_test=pd.read_csv('data/3-processed_data/Y_test.csv')
#Loading model
rnn_model = load_model('models/rnn_model.h5')

#lets try it on two different comments we manually type
#com='I dont give a shit'
com='thank you for your feedback'

com = word_tokenize(com)
tokenizer.fit_on_texts(com)
com = tokenizer.texts_to_sequences(com)
com=list(itertools.chain.from_iterable(com))
com = pad_sequences([com], padding='post', maxlen=200)
r=np.reshape(com[0], newshape=(1,200))
pred=rnn_model.predict(r, batch_size=1)
labels = list(Y_test.columns)

a=list(pred[0])
results=[]
for idx,val in enumerate(a):
    if val>0.5:
        results.append(labels[idx])
        


print('The model classifies your comment as ',str(results),'.')

## Recommendations
* Ban most common words in toxic comments
* Use Recurrent neural network model to create alerts on critical commenths like threats
* Provide more data to improve performance

## Future Work

* Try pretrained models
* Focus on threat comments
* Train for more ecpoch on more powerful machine
* Deploy on AWS
